Update 3
Added post route for user registration.  The registration checks if the user already exists in the db and if not it creates a new user and attaches the registration form fields to the user model.  It then hashes the input password and adds the hashed password back into the user object.
Added passport.js file that contains passport strategy and collects payload from the registration page.  It then decodes the web token payload.  The payload user id is verified against the users in the db.  If the user exists and is verified the user is returned.