Update 3
Added login route and code.  The login function takes the form fields email and password and first checks if the user exists in the db.  It then checks the password against the one in the database and if it matches creates a jwt payload.  The payload is signed and a token is returned.