Update 6
Added user login component.  The component post to the login endpoint and takes an email and password.  The email and password are then searched for in the database and if they exist a success message is displayed along with the jwt token.  Added a loading component to display a loading message while data is being fetched.