Update 11
Added post route endpoint for Creating new posts.  When an id and token are passed to the route the user is looked up by the input id.  The input token is then compared against the user's token which should be there if they are logged in.  If the tokens match the new post is created using the post collection and the user id and post description are added.  The new post is saved to the db and also added to the current user's post array.