Update 22
Added count field and user name fields to post schema.  Created route controller for adding count to a post.  The controller searchers for the post by its id then increments the count value, saves and returns the updated count.  Began working on the addCount method and adding functionality to the displayPosts component to allow incrementing on button click within a specific post frame.