Update 23
Began work adding state and fetch information to addCount component.  The component takes the clicked on count button and submits the postId and count to the controller and should return the updated post and update the post list.  Working on resolving an issue with submitting the postId correctly to the controller.