Update 25
Began work on a display single post component.  The component will display a new screen when a post on the post list is clicked.  The new screen will contain the post title and body.  It will also contain an input field for comments and a list of comments related to that post below.  Created the basic rendering functionality and api call methods.  Completed some of the error handling and validation.  The component will check if the user is logged in when a new comment is created and if not an error message is displayed.