Update 27
Fixed issue with post prop not transferring properly to the singlePost component.  Extracted the rendering to a separate method within the component.  Began work on styling within the single post and post list components.  Next steps will be to add a logout button that propogates the empty token state down to all the child components and renders the login component again.