Update 33
Fixed an issue where getCount endpoint wasn't getting the new logged in userId.  The clicked on post id was being fed to the component instead of current user id.  Fixed an issue with component not showing correctly after a user fails to login.  Added conditional rendering to the main app to display register/login components if noone has logged in yet.  Once a user logs in it will display the userIsLoggedComponent or else display any login errors.