Update 37
Updated endpoint for create comment to now correctly check the user is logged in then create the new comment for the user.  Created new component on the frontend that includes a comment form, calls the api and displays the returned comment data.  Next steps to create a list of comments and work on count modification and filtering for the list.