Update 38
Refactored unnecessary code.  Removed code where props were added to state as it was causing issues with props updating properly on re-render.  Updated any places where state was used instead of props.  Saved props into a separate obj in commentComp.  The obj is then passed to the fetch function for passing to the comment endpoint.