Update 40
Extracted fetch function to parent singlePost component.  Shared props for the get comment function directly to the commentList and CreateComment components.  Added updating to the createComment component so it now updates the CommentList component after a new comment is created and refreshes the list.  Comment creation and list is now working.  Next steps are to add comment count functionality.