Update 41
Created endpoint for adding/subtracting count to a comment.  Created logic for updating the comment count and making sure a single user can't add more or less than 1 point to a comment.  Added logic for keeping track of users who've counted and verifying they are logged in before being able to add to the comment count.  Next steps are to add functionality for viewing next 10 posts or comments in the list with buttons for previous and next etc.