Update 53
Added some css to comment height and spacing.  Fixed a bug that was causing pasted in code in textbox to get rendered.  Added html sanitization library to parse and include only certain tags in the html.  Fixed styling issues with richTextEditor and removed unnecessary buttons and dropdowns from the component.  Need to refactor the configuration out of the handleSubmit component.