Update 57
Refactored code in updateToken method to only take one param obj and update the state instead of requiring multiple params and then setting the state with them.  Added currentUserId to deleteCommentComp and included passing the user id into the delete comment endpoint.  Updated post and comment endpoints to check against the super user id and if they don't match the comment/post id or the super user id to throw an error.  Super user functionality now complete to allow a super user to delete any comments and posts.  Also, added logic in the logout  component to render a super user tag if the super user is logged in.